Update and rename matrixmultiplication.py to matrix_multiplication.py
